requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
Flask==1.1.2
Flask-Compress==1.4.0
Flask-SQLAlchemy==2.4.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
logparser==0.8.2
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.2
parsel==1.5.2
pexpect==4.8.0
Protego==0.1.16
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycodestyle==2.5.0
pycountry==19.8.18
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.5.0
requests==2.23.0
Scrapy==2.0.1
scrapyd==1.2.1
scrapyd-client==1.2.0a1
scrapydweb==1.4.0
service-identity==18.1.0
six==1.14.0
SQLAlchemy==1.3.15
Twisted==20.3.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
w3lib==1.21.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
zope.interface==5.0.2
